files naming & login bugfix
