--- lib/index.ts
@@ -1,15 +1,0 @@
-export { default as UserQuery } from './queries/UserQuery';
-export { default as Login } from './queries/login';
-export { default as CreateUser } from './mutations/createUser';
-export { default as DeleteUser } from './mutations/deleteUser';
-export { default as DeletePost } from './mutations/deletePost';
-export { default as Follow } from './mutations/follow';
-export { default as Unfollow } from './mutations/unfollow';
-export { default as Like } from './mutations/like';
-export { default as CreateCollection } from './mutations/createCollection';
-export { default as CreatePost } from './mutations/createPost';
-export { default as EditProfile } from './mutations/editProfile';
-export { default as GetPaths } from './queries/getPaths';
-export { default as GetTagInfo } from './queries/getTagInfo';
-export { default as SuggestedUsersQuery } from './queries/SuggestedUsersQuery';
-export { default as GetMapPlaces } from './queries/GetMapPlaces';
