+++ lib/mutations/index.ts
@@ -1,0 +1,9 @@
+export { default as CreateUser } from './createUser';
+export { default as DeleteUser } from './deleteUser';
+export { default as DeletePost } from './deletePost';
+export { default as Follow } from './follow';
+export { default as Unfollow } from './unfollow';
+export { default as Like } from './like';
+export { default as CreateCollection } from './createCollection';
+export { default as CreatePost } from './createPost';
+export { default as EditProfile } from './editProfile';
