+++ lib/queries/index.ts
@@ -1,0 +1,6 @@
+export { default as UserQuery } from './UserQuery';
+export { default as Login } from './login';
+export { default as GetPaths } from './getPaths';
+export { default as GetTagInfo } from './getTagInfo';
+export { default as SuggestedUsersQuery } from './SuggestedUsersQuery';
+export { default as FilterPlacesQuery } from './FilterPlaces';
